requirements_dev.txt
black==23.3.0                                     # must match .pre-commit-config.yaml
flake8==6.0.0; python_version >= "3.8"            # must match .pre-commit-config.yaml
flake8-bugbear==23.3.23; python_version >= "3.8"  # must match .pre-commit-config.yaml
flake8-noqa==1.3.1; python_version >= "3.8"       # must match .pre-commit-config.yaml
isort==5.12.0; python_version >= "3.8"            # must match .pre-commit-config.yaml
pre-commit>=3.2.2
pre-commit-hooks==4.4.0
trio-typing[mypy]>=0.8.0
types-typed-ast>=1.5.8.6
